+++ addons/TBMod_ZeusModuls/ReadMe.md
@@ -1,0 +1,4 @@
+# TBMod Main
+
+MainDeveloper: [shukari](https://github.com/shukari)
+Developer:     [Eric Ruhland](https://github.com/Er1807)
