+++ addons/TBMod_ZeusModuls/configs/CfgFunctions.hpp
@@ -1,0 +1,23 @@
+﻿/*
+    Part of the TBMod ( https://github.com/shukari/TBMod )
+    Developed by http://tacticalbacon.de
+*/
+class CfgFunctions
+{
+    class TBMod_main
+    {
+        tag = "TB";
+
+        class functions
+        {
+            file = "\TBMod_ZeusModules\functions";
+            class moduleNachschubAbwurf {};
+            class moduleDisablePath {};
+            class modulePrioSichtbarkeit {};
+            class moduleHandleAsPlayerMedical {};
+            class moduleApplyLoadout {};
+            class moduleFrontverlauf {};
+        };
+
+    };
+};
