+++ addons/TBMod_ZeusModuls/functions/fn_moduleDisablePath.sqf
