+++ addons/TBMod_main/configs/CfgFunctions.hpp
@@ -20,9 +20,6 @@ class CfgFunctions
             class modifierFOBs {};
             class displayName {};
             class isTBMission {};
-            class moduleNachschubAbwurf {};
-            class moduleDisablePath {};
-            class modulePrioSichtbarkeit {};
             class insertChildrenGruppen {};
             class garage {};
             class garageGUI {};
@@ -38,9 +35,6 @@ class CfgFunctions
             class sitAI {};
             class animationAI {};
             class keysForAI {};
-            class moduleHandleAsPlayerMedical {};
-            class moduleApplyLoadout {};
-            class moduleFrontverlauf {};
             class debug {};
             class forceRagdoll {};
             class attachToRelative {};
